# api-specs/openapi.yaml
openapi: 3.1.0
info:
  title: VetPrescription API
  version: 0.0.0
  description: >
    Veterinary prescription generator REST API.
    Compliant with Article 105.5 of Regulation (EU) 2019/6 (Catalonia jurisdiction).
  contact:
    name: Vet Prescription
servers:
  - url: http://localhost:8080/api
    description: Local Docker Compose

tags:
  - name: Prescriptions
    description: Veterinary prescription management and PDF generation
  - name: Vets
    description: Veterinarian profile management

paths:
  /prescriptions:
    post:
      operationId: createPrescription
      tags: [Prescriptions]
      summary: Create a new veterinary prescription
      description: >
        Creates a prescription record compliant with Article 105.5 of
        Regulation (EU) 2019/6. All mandatory fields must be provided.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePrescriptionRequest'
            example:
              vet:
                name: "Dr. Joan Puig Domènech"
                licenceNumber: "CAT-1234"
                clinicName: "Clínica Veterinària Puig"
                address: "Carrer Major 10, 08001 Barcelona"
                phone: "+34 93 123 45 67"
                email: "joan.puig@clinicapuig.cat"
              owner:
                name: "Maria Soler"
                address: "Avinguda Diagonal 200, 08021 Barcelona"
                phone: "+34 612 345 678"
                cifDni: "12345678A"
              patient:
                animalName: "Rex"
                species: "Canis lupus familiaris"
                breed: "Pastor Alemany"
              items:
                - drugName: "Amoxicil·lina"
                  quantity: "1 envàs (14 comprimits)"
                  pharmaceuticalForm: "Comprimits"
                  dosageRegimen: "1 comprimit cada 12h durant 7 dies"
                  withdrawalPeriod: null
              warnings: "Conservar en lloc fresc i sec."
              isOffLabel: false
              isAntimicrobialSpecialUse: false
      responses:
        '201':
          description: Prescription created successfully
          headers:
            Location:
              description: URL of the created prescription
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrescriptionResponse'
        '400':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://tools.ietf.org/html/rfc7807"
                title: "Validation Failed"
                status: 400
                detail: "One or more validation errors occurred."
                errors:
                  "vet.name": ["'Vet Name' must not be empty."]
                  "items": ["At least one prescription item is required."]
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  /prescriptions/{id}:
    get:
      operationId: getPrescriptionById
      tags: [Prescriptions]
      summary: Retrieve a prescription by ID
      description: Returns the full prescription record for display in the UI.
      parameters:
        - name: id
          in: path
          required: true
          description: MongoDB ObjectId of the prescription
          schema:
            type: string
            pattern: '^[a-fA-F0-9]{24}$'
          example: "6647a2f3e4b0c1d2e3f40001"
      responses:
        '200':
          description: Full prescription record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrescriptionDetailResponse'
              example:
                id: "6647a2f3e4b0c1d2e3f40001"
                prescriptionNumber: "RX-2026-0001"
                date: "2026-05-17"
                vet:
                  name: "Dr. Joan Puig Domènech"
                  licenceNumber: "CAT-1234"
                  clinicName: "Clínica Veterinària Puig"
                  address: "Carrer Major 10, 08001 Barcelona"
                  phone: "+34 93 123 45 67"
                  email: "joan.puig@clinicapuig.cat"
                owner:
                  name: "Maria Soler"
                  address: "Avinguda Diagonal 200, 08021 Barcelona"
                  phone: "+34 612 345 678"
                  cifDni: "12345678A"
                patient:
                  animalName: "Rex"
                  species: "Canis lupus familiaris"
                  breed: "Pastor Alemany"
                items:
                  - drugName: "Amoxicil·lina"
                    quantity: "1 envàs (14 comprimits)"
                    pharmaceuticalForm: "Comprimits"
                    dosageRegimen: "1 comprimit cada 12h durant 7 dies"
                    withdrawalPeriod: null
                warnings: "Conservar en lloc fresc i sec."
                isOffLabel: false
                isAntimicrobialSpecialUse: false
        '404':
          description: Prescription not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://tools.ietf.org/html/rfc7807"
                title: "Not Found"
                status: 404
                detail: "Prescription '6647a2f3e4b0c1d2e3f40001' was not found."
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  /prescriptions/{id}/pdf:
    post:
      operationId: generatePrescriptionPdf
      tags: [Prescriptions]
      summary: Generate and store the PDF for a prescription
      description: >
        Generates a standards-compliant PDF document formatted according to
        the Catalan/Spanish veterinary prescription layout (Art. 105.5,
        Regulation EU 2019/6) and saves it to storage. Idempotent: calling
        it again overwrites the previously stored file.
      parameters:
        - name: id
          in: path
          required: true
          description: MongoDB ObjectId of the prescription
          schema:
            type: string
            pattern: '^[a-fA-F0-9]{24}$'
          example: "6647a2f3e4b0c1d2e3f40001"
      responses:
        '201':
          description: PDF generated and stored successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdfResponse'
              example:
                url: "http://localhost:8080/pdfs/RX-2026-0001.pdf"
                prescriptionNumber: "RX-2026-0001"
        '404':
          description: Prescription not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://tools.ietf.org/html/rfc7807"
                title: "Not Found"
                status: 404
                detail: "Prescription '6647a2f3e4b0c1d2e3f40001' was not found."
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

    get:
      operationId: getPrescriptionPdfUrl
      tags: [Prescriptions]
      summary: Retrieve the URL of the stored PDF for a prescription
      description: >
        Returns the URL of the previously generated PDF. Returns 404 if
        the PDF has not been generated yet for this prescription.
      parameters:
        - name: id
          in: path
          required: true
          description: MongoDB ObjectId of the prescription
          schema:
            type: string
            pattern: '^[a-fA-F0-9]{24}$'
          example: "6647a2f3e4b0c1d2e3f40001"
      responses:
        '200':
          description: PDF URL retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdfResponse'
              example:
                url: "http://localhost:8080/pdfs/RX-2026-0001.pdf"
                prescriptionNumber: "RX-2026-0001"
        '404':
          description: Prescription or PDF not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://tools.ietf.org/html/rfc7807"
                title: "Not Found"
                status: 404
                detail: "No PDF has been generated yet for prescription '6647a2f3e4b0c1d2e3f40001'."
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  /vets/profile:
    post:
      operationId: saveVetProfile
      tags: [Vets]
      summary: Save or update the veterinarian profile
      description: >
        Upserts the veterinarian profile. The profile is used to pre-fill
        the prescription form on subsequent visits.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VetProfileRequest'
            example:
              name: "Dr. Joan Puig Domènech"
              licenceNumber: "CAT-1234"
              clinicName: "Clínica Veterinària Puig"
              address: "Carrer Major 10, 08001 Barcelona"
              phone: "+34 93 123 45 67"
              email: "joan.puig@clinicapuig.cat"
      responses:
        '200':
          description: Profile saved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VetProfileResponse'
              example:
                name: "Dr. Joan Puig Domènech"
                licenceNumber: "CAT-1234"
                clinicName: "Clínica Veterinària Puig"
                address: "Carrer Major 10, 08001 Barcelona"
                phone: "+34 93 123 45 67"
                email: "joan.puig@clinicapuig.cat"
        '400':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://tools.ietf.org/html/rfc7807"
                title: "Validation Failed"
                status: 400
                detail: "One or more validation errors occurred."
                errors:
                  "name": ["'Name' must not be empty."]
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

    get:
      operationId: getVetProfile
      tags: [Vets]
      summary: Retrieve the saved veterinarian profile
      description: >
        Returns the veterinarian profile previously saved via POST /vets/profile.
        Returns 404 if no profile has been saved yet.
      responses:
        '200':
          description: Vet profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VetProfileResponse'
              example:
                name: "Dr. Joan Puig Domènech"
                licenceNumber: "CAT-1234"
                clinicName: "Clínica Veterinària Puig"
                address: "Carrer Major 10, 08001 Barcelona"
                phone: "+34 93 123 45 67"
                email: "joan.puig@clinicapuig.cat"
        '404':
          description: No profile saved yet
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://tools.ietf.org/html/rfc7807"
                title: "Not Found"
                status: 404
                detail: "No vet profile has been saved yet."
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

components:
  schemas:
    # ── Shared value objects ───────────────────────────────────────────────

    VetDto:
      type: object
      description: >
        Veterinarian details — Art. 105.5(d): full name and contact details
        including professional licence number.
      required: [name, licenceNumber, clinicName, address, phone, email]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          description: Full name of the veterinarian
        licenceNumber:
          type: string
          minLength: 1
          maxLength: 50
          description: Professional licence number (col·legi veterinari)
        clinicName:
          type: string
          minLength: 1
          maxLength: 200
          description: Name of the clinic or practice
        address:
          type: string
          minLength: 1
          maxLength: 300
          description: Clinic address
        phone:
          type: string
          minLength: 1
          maxLength: 30
          description: Clinic or vet contact phone
        email:
          type: string
          format: email
          description: Vet contact email

    OwnerDto:
      type: object
      description: >
        Animal owner details — Art. 105.5(b): full name and contact details
        of the owner or responsible person.
      required: [name, address, phone, cifDni]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          description: Full name of the owner
        address:
          type: string
          minLength: 1
          maxLength: 300
          description: Owner address
        phone:
          type: string
          minLength: 1
          maxLength: 30
          description: Owner phone
        cifDni:
          type: string
          minLength: 1
          maxLength: 20
          description: Owner tax ID (CIF or DNI)

    PatientDto:
      type: object
      description: >
        Animal identification — Art. 105.5(a): identification of the animal
        or group of animals subject to treatment.
      required: [animalName, species, breed]
      properties:
        animalName:
          type: string
          minLength: 1
          maxLength: 100
          description: Name or identifier of the animal
        species:
          type: string
          minLength: 1
          maxLength: 100
          description: Species (e.g. Canis lupus familiaris)
        breed:
          type: string
          minLength: 1
          maxLength: 100
          description: Breed

    PrescriptionItemDto:
      type: object
      description: >
        A single prescribed medicinal product — Art. 105.5(f)(g)(h)(i):
        drug name, pharmaceutical form, quantity, and dosage regimen.
      required: [drugName, quantity, pharmaceuticalForm, dosageRegimen]
      properties:
        drugName:
          type: string
          minLength: 1
          maxLength: 200
          description: Name of the prescribed medicinal product
        quantity:
          type: string
          minLength: 1
          maxLength: 100
          description: Quantity or number of units prescribed — Art. 105.5(h)
        pharmaceuticalForm:
          type: string
          minLength: 1
          maxLength: 100
          description: Pharmaceutical form (e.g. comprimits, solució oral) — Art. 105.5(g)
        dosageRegimen:
          type: string
          minLength: 1
          maxLength: 300
          description: >
            Dosage regimen: dose, frequency, and duration — Art. 105.5(i)
        withdrawalPeriod:
          type: string
          nullable: true
          maxLength: 100
          description: >
            Withdrawal period for food-producing animals, or null for
            non-food-producing species — Art. 105.5(j)

    # ── Request bodies ─────────────────────────────────────────────────────

    CreatePrescriptionRequest:
      type: object
      description: >
        Full prescription data. All Art. 105.5 mandatory fields are required.
        Art. 105.5(c) date and (e) signature are set server-side.
      required: [vet, owner, patient, items]
      properties:
        vet:
          $ref: '#/components/schemas/VetDto'
        owner:
          $ref: '#/components/schemas/OwnerDto'
        patient:
          $ref: '#/components/schemas/PatientDto'
        items:
          type: array
          minItems: 1
          description: List of prescribed medicinal products (at least one required)
          items:
            $ref: '#/components/schemas/PrescriptionItemDto'
        warnings:
          type: string
          nullable: true
          maxLength: 1000
          description: >
            Any warnings for correct/prudent use — Art. 105.5(k)
        isOffLabel:
          type: boolean
          default: false
          description: >
            True if prescribed under off-label cascade
            (Art. 112/113/114, Regulation EU 2019/6) — Art. 105.5(l)
        isAntimicrobialSpecialUse:
          type: boolean
          default: false
          description: >
            True if antimicrobial prescribed for prophylaxis or metaphylaxis
            (Art. 107.3/4, Regulation EU 2019/6) — Art. 105.5(m)

    # ── Response bodies ────────────────────────────────────────────────────

    PrescriptionResponse:
      type: object
      description: >
        Minimal response after prescription creation. The caller already
        holds all other data from the request body.
      required: [id, prescriptionNumber, date]
      properties:
        id:
          type: string
          description: MongoDB ObjectId — use to download the PDF
          example: "6647a2f3e4b0c1d2e3f40001"
        prescriptionNumber:
          type: string
          description: Human-readable prescription reference
          example: "RX-2026-0001"
        date:
          type: string
          format: date
          description: Date of issue assigned by the server — Art. 105.5(c)
          example: "2026-05-17"

    PrescriptionDetailResponse:
      type: object
      description: Full prescription record for display in the UI
      required: [id, prescriptionNumber, date, vet, owner, patient, items, isOffLabel, isAntimicrobialSpecialUse]
      properties:
        id:
          type: string
          example: "6647a2f3e4b0c1d2e3f40001"
        prescriptionNumber:
          type: string
          example: "RX-2026-0001"
        date:
          type: string
          format: date
          example: "2026-05-17"
        vet:
          $ref: '#/components/schemas/VetDto'
        owner:
          $ref: '#/components/schemas/OwnerDto'
        patient:
          $ref: '#/components/schemas/PatientDto'
        items:
          type: array
          items:
            $ref: '#/components/schemas/PrescriptionItemDto'
        warnings:
          type: string
          nullable: true
        isOffLabel:
          type: boolean
        isAntimicrobialSpecialUse:
          type: boolean

    PdfResponse:
      type: object
      description: >
        Result of a PDF generation request. The url points directly to the
        stored PDF file and can be opened or downloaded by the browser.
      required: [url, prescriptionNumber]
      properties:
        url:
          type: string
          format: uri
          description: Direct URL to the generated PDF file
          example: "http://localhost:8080/pdfs/RX-2026-0001.pdf"
        prescriptionNumber:
          type: string
          description: Human-readable prescription reference
          example: "RX-2026-0001"

    # ── Vet Profile ────────────────────────────────────────────────────────

    VetProfileRequest:
      type: object
      description: Veterinarian profile data for saving/updating
      required: [name, licenceNumber, clinicName, address, phone, email]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        licenceNumber:
          type: string
          minLength: 1
          maxLength: 50
        clinicName:
          type: string
          minLength: 1
          maxLength: 200
        address:
          type: string
          minLength: 1
          maxLength: 300
        phone:
          type: string
          minLength: 1
          maxLength: 30
        email:
          type: string
          format: email

    VetProfileResponse:
      type: object
      description: Saved veterinarian profile
      required: [name, licenceNumber, clinicName, address, phone, email]
      properties:
        name:
          type: string
        licenceNumber:
          type: string
        clinicName:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email

    # ── Error ──────────────────────────────────────────────────────────────

    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string